+++ tests/test_components.py
@@ -643,7 +643,9 @@ class TestGfsDycore(ComponentBase):
         component = self.get_component_instance()
         state = climt.get_default_state(
             [component], x=component.grid_definition['x'],
-            y=component.grid_definition['y'], z=component.grid_definition['mid_levels'])
+            y=component.grid_definition['y'],
+            mid_levels=component.grid_definition['mid_levels'],
+            interface_levels=component.grid_definition['interface_levels'])
 
         dcmip = climt.DcmipInitialConditions()
         out = dcmip(state, add_perturbation=True)